Replace old level module with new one
